max-arg-count: turn max-arg-count into int again
To avoid potential casting/conversion issues between different integer
types (as `argc` is int), convert the value read from `max-arg-count`
into int when parsing it.

+++ src/firejail/firejail.h
@@ -878,7 +878,7 @@ extern char *xvfb_screen;
 extern char *xvfb_extra_params;
 extern char *netfilter_default;
 extern unsigned long join_timeout;
-extern long max_arg_count;
+extern int max_arg_count;
 extern unsigned long max_arg_len;
 extern char *config_seccomp_error_action_str;
 extern char *config_seccomp_filter_add;
